[skipruntime] Add @skipruntime/runtime package
- Manage load of @skipruntime/addon if available
