+++ examples/hackernews/reactive_service/package.json
@@ -10,7 +10,7 @@
   "description": "",
   "dependencies": {
     "postgres": "^3.4.5",
-    "@skipruntime/wasm": "^0.0.1",
+    "@skipruntime/runtime": "^0.0.1",
     "@skipruntime/server": "^0.0.4"
   },
   "devDependencies": {
