+++ package.json
@@ -13,6 +13,7 @@
     "skipruntime-ts/tests",
     "skipruntime-ts/wasm",
     "skipruntime-ts/addon",
+    "skipruntime-ts/runtime",
     "skipruntime-ts/server",
     "skipruntime-ts/examples",
     "sql/ts",
