+++ skipruntime-ts/helpers/package.json
@@ -4,7 +4,6 @@
   "type": "module",
   "exports": {
     ".": "./dist/index.js",
-    "./errors.js": "./dist/errors.js",
     "./rest.js": "./dist/rest.js",
     "./external.js": "./dist/external.js"
   },
