+++ skipruntime-ts/runtime/eslint.config.js
@@ -1,0 +1,3 @@
+import config from "@skiplabs/eslint-config";
+
+export default [...config];
