+++ skipruntime-ts/runtime/package.json
@@ -1,0 +1,20 @@
+{
+  "name": "@skipruntime/runtime",
+  "version": "0.0.1",
+  "type": "module",
+  "exports": {
+    ".": "./dist/index.js"
+  },
+  "scripts": {
+    "build": "tsc",
+    "clean": "rm -rf dist",
+    "lint": "eslint"
+  },
+  "dependencies": {
+    "@skipruntime/helpers": "^0.0.4",
+    "@skipruntime/wasm": "^0.0.1"
+  },
+  "optionalDependencies": {
+    "@skipruntime/addon": "^0.0.1"
+  }
+}
