+++ skipruntime-ts/server/package.json
@@ -15,6 +15,6 @@
   },
   "dependencies": {
     "express": "^4.21.1",
-    "@skipruntime/wasm": "^0.0.1"
+    "@skipruntime/runtime": "^0.0.1"
   }
 }
