+++ skipruntime-ts/wasm/package.json
@@ -20,7 +20,6 @@
     "@skiplang/json": "^0.0.1",
     "@skipruntime/api": "^0.0.3",
     "@skipruntime/core": "^0.0.1",
-    "@skipruntime/helpers": "^0.0.4",
     "@skip-wasm/json": "^1.0.3",
     "@skip-wasm/std": "^1.0.2"
   }
